# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Go Web API Boilerplate
  version: 1.0.0
  description: OpenAPI specification for the v1 HTTP API.
servers:
  - url: http://localhost:8080/api/v1
...
paths:
  /tenants:
    post:
      tags: [Tenants]
      summary: Create tenant
      operationId: createTenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
      responses:
        '201':
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTenantResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
  /branches:
    post:
      tags: [Tenants]
      summary: Create branch
      operationId: createBranch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBranchRequest'
      responses:
        '201':
          description: Branch created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateBranchResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
  /subscriptions:
    post:
      tags: [Tenants]
      summary: Create subscription
      operationId: createSubscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriptionRequest'
      responses:
        '201':
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateSubscriptionResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: PASETO
  responses:
    ErrorResponse:
      description: API error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    CreateTenantRequest:
      type: object
      required: [name, slug, domain, type]
      properties:
        name:
          type: string
        slug:
          type: string
        domain:
          type: string
        type:
          type: string
          enum: [bakery, pharmacy, restaurant]
    CreateTenantResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        domain:
          type: string
        type:
          type: string
        created_at:
          type: string
          format: date-time
    CreateBranchRequest:
      type: object
      required: [tenant_id, name, code]
      properties:
        tenant_id:
          type: string
          format: uuid
        name:
          type: string
        code:
          type: string
    CreateBranchResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        name:
          type: string
        code:
          type: string
    CreateSubscriptionRequest:
      type: object
      required: [tenant_id, plan, status]
      properties:
        tenant_id:
          type: string
          format: uuid
        plan:
          type: string
        status:
          type: string
          enum: [trial, active, past_due, suspended, canceled]
        starts_at:
          type: string
          format: date-time
        ends_at:
          type: string
          format: date-time
    CreateSubscriptionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        plan:
          type: string
        status:
          type: string
        starts_at:
          type: string
          format: date-time
        ends_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: auth/unauthorized
        errors:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
            - type: object
              additionalProperties: true