# generated/typescript-fetch/models/CreateBranchRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Go Web API Boilerplate
 * OpenAPI specification for the v1 HTTP API.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateBranchRequest
 */
export interface CreateBranchRequest {
    /**
     * 
     * @type {string}
     * @memberof CreateBranchRequest
     */
    tenantId: string;
    /**
     * 
     * @type {string}
     * @memberof CreateBranchRequest
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof CreateBranchRequest
     */
    code: string;
}

/**
 * Check if a given object implements the CreateBranchRequest interface.
 */
export function instanceOfCreateBranchRequest(value: object): value is CreateBranchRequest {
    if (!('tenantId' in value) || value['tenantId'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('code' in value) || value['code'] === undefined) return false;
    return true;
}

export function CreateBranchRequestFromJSON(json: any): CreateBranchRequest {
    return CreateBranchRequestFromJSONTyped(json, false);
}

export function CreateBranchRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateBranchRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'tenantId': json['tenant_id'],
        'name': json['name'],
        'code': json['code'],
    };
}

export function CreateBranchRequestToJSON(json: any): CreateBranchRequest {
    return CreateBranchRequestToJSONTyped(json, false);
}

export function CreateBranchRequestToJSONTyped(value?: CreateBranchRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'tenant_id': value['tenantId'],
        'name': value['name'],
        'code': value['code'],
    };
}