# generated/typescript-fetch/models/CreateBranchResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Go Web API Boilerplate
 * OpenAPI specification for the v1 HTTP API.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateBranchResponse
 */
export interface CreateBranchResponse {
    /**
     * 
     * @type {string}
     * @memberof CreateBranchResponse
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateBranchResponse
     */
    tenantId?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateBranchResponse
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateBranchResponse
     */
    code?: string;
}

/**
 * Check if a given object implements the CreateBranchResponse interface.
 */
export function instanceOfCreateBranchResponse(value: object): value is CreateBranchResponse {
    return true;
}

export function CreateBranchResponseFromJSON(json: any): CreateBranchResponse {
    return CreateBranchResponseFromJSONTyped(json, false);
}

export function CreateBranchResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateBranchResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'tenantId': json['tenant_id'] == null ? undefined : json['tenant_id'],
        'name': json['name'] == null ? undefined : json['name'],
        'code': json['code'] == null ? undefined : json['code'],
    };
}

export function CreateBranchResponseToJSON(json: any): CreateBranchResponse {
    return CreateBranchResponseToJSONTyped(json, false);
}

export function CreateBranchResponseToJSONTyped(value?: CreateBranchResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'tenant_id': value['tenantId'],
        'name': value['name'],
        'code': value['code'],
    };
}