# generated/typescript-fetch/models/CreateTenantResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Go Web API Boilerplate
 * OpenAPI specification for the v1 HTTP API.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateTenantResponse
 */
export interface CreateTenantResponse {
    /**
     * 
     * @type {string}
     * @memberof CreateTenantResponse
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateTenantResponse
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateTenantResponse
     */
    slug?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateTenantResponse
     */
    domain?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateTenantResponse
     */
    type?: string;
    /**
     * 
     * @type {Date}
     * @memberof CreateTenantResponse
     */
    createdAt?: Date;
}

/**
 * Check if a given object implements the CreateTenantResponse interface.
 */
export function instanceOfCreateTenantResponse(value: object): value is CreateTenantResponse {
    return true;
}

export function CreateTenantResponseFromJSON(json: any): CreateTenantResponse {
    return CreateTenantResponseFromJSONTyped(json, false);
}

export function CreateTenantResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateTenantResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'slug': json['slug'] == null ? undefined : json['slug'],
        'domain': json['domain'] == null ? undefined : json['domain'],
        'type': json['type'] == null ? undefined : json['type'],
        'createdAt': json['created_at'] == null ? undefined : (new Date(json['created_at'])),
    };
}

export function CreateTenantResponseToJSON(json: any): CreateTenantResponse {
    return CreateTenantResponseToJSONTyped(json, false);
}

export function CreateTenantResponseToJSONTyped(value?: CreateTenantResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'slug': value['slug'],
        'domain': value['domain'],
        'type': value['type'],
        'created_at': value['createdAt'] == null ? value['createdAt'] : value['createdAt'].toISOString(),
    };
}